model_runtime/requirements.txt
fastapi
uvicorn[standard]
pillow
requests
modelscope
accelerate
torch
torchvision
torchaudio
git+https://github.com/huggingface/transformers
qwen-vl-utils[decord]==0.0.8